+++ pkg/cmdsetup/data.go
@@ -45,11 +45,12 @@ func newDataCmd(runners ...run.Unit) *cobra.Command {
 	}
 	pipeline := sub.NewServer()
 	localPipeline := queue.Local()
-	streamSvc, err := stream.NewService(ctx, metaSvc, pipeline)
+	metricSvc := observability.NewMetricService(metaSvc, localPipeline, "data", nil)
+	streamSvc, err := stream.NewService(ctx, metaSvc, pipeline, metricSvc)
 	if err != nil {
 		l.Fatal().Err(err).Msg("failed to initiate stream service")
 	}
-	measureSvc, err := measure.NewService(ctx, metaSvc, pipeline, localPipeline)
+	measureSvc, err := measure.NewService(ctx, metaSvc, pipeline, localPipeline, metricSvc)
 	if err != nil {
 		l.Fatal().Err(err).Msg("failed to initiate measure service")
 	}
@@ -58,7 +59,6 @@ func newDataCmd(runners ...run.Unit) *cobra.Command {
 		l.Fatal().Err(err).Msg("failed to initiate query processor")
 	}
 	profSvc := observability.NewProfService()
-	metricSvc := observability.NewMetricService(metaSvc, localPipeline, "data", nil)
 
 	var units []run.Unit
 	units = append(units, runners...)
@@ -66,14 +66,12 @@ func newDataCmd(runners ...run.Unit) *cobra.Command {
 		metaSvc,
 		localPipeline,
 		pipeline,
+		metricSvc,
 		measureSvc,
 		streamSvc,
 		q,
 		profSvc,
 	)
-	if metricSvc != nil {
-		units = append(units, metricSvc)
-	}
 	dataGroup := run.NewGroup("data")
 	dataGroup.Register(units...)
 	dataCmd := &cobra.Command{